refactor(multisite): change namespace of multisite models
